Computer-Vision/Detect-Object-In-Image/requirements.txt
python-dotenv~=1.0.0
azure-identity>=1.15.0
azure-cognitiveservices-vision-customvision
azure-ai-vision-imageanalysis==1.0.0
matplotlib